requirements.txt
typer[all]^=0.6.1
pyserial^=3.5
